pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graph_ragu"
version = "1.0.4"
description = "GraphRAG system"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.22.0",
    "aiolimiter>=1.2.1",
    "apscheduler>=3.11.1",
    "gspread>=6.2.1",
    "google-auth>=2.41.1",
    "graspologic>=3.4.4",
    "httpx>=0.28.1",
    "instructor>=1.11.3",
    "loguru>=0.7.3",
    "nano-vectordb>=0.0.4.3",
    "networkx>=3.4.2",
    "neo4j>=5.28.1",
    "nltk>=3.9.2",
    "numpy>=1.26.4",
    "openai>=1.109.1",
    "pandas>=2.3.3",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.11.0",
    "razdel>=0.5.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.11.0",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
local = [
    "sentence-transformers>=5.1.1",
    "transformers>=4.57.0",
    "smart_chunker @ git+https://github.com/AsphodelRem/smart_chunker.git",
]

test = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools.packages.find]
include = ["ragu*", "apps*"]

[tool.coverage.run]
omit = [
    "tests/*",
]

[tool.coverage.report]
skip_empty = true
ignore_errors = true  # torch generates files in /tmp/ dir that crash `coverage report -m` command
